+++ spec/rook/jetty_async_spec.clj
@@ -13,34 +13,16 @@
     [io.aviso.rook
      [async :as async]
      [utils :as utils]
-     [jetty-async-adapter :as jetty]]
-    [clojure.tools.logging :as l]))
+     [clj-http :as ch]
+     [jetty-async-adapter :as jetty]]))
 
 (describe "io.aviso.rook.jetty-async-adapter"
 
+  (with-all handler
+            (ch/handler "http://localhost:9988/"))
+
   (with-all server
-            (let [loopback-handler (fn [request]
-                                     #_ (l/info "loopback:" (utils/pretty-print request))
-                                     (let [body-params (:body-params request)
-                                           body (if body-params (pr-str body-params))
-                                           client-request (-> request
-                                                              ;; Need to convert :body-params to :body
-                                                              ;; and the :content-type :edn will take care of sending an EDN stream
-                                                              ;; over the wire.
-                                                              (dissoc :body-params :params)
-                                                              (assoc :url (str "http://localhost:9988/" (:uri request))
-                                                                     :body body
-                                                                     :content-type :edn
-                                                                     :accept :edn
-                                                                     :follow-redirects false
-                                                                     :throw-exceptions false
-                                                                     :as :clojure))]
-                                       (thread
-                                         ;; I'd love it if there was a way to do this asynchronously via some kind of callback. We're using
-                                         ;; a thread from the core.async pool AND a thread from the connection manager.
-                                         (client/request client-request))))]
-              (->
-                (rook/namespace-handler {:async?        true
+            (-> (rook/namespace-handler {:async?        true
                                          :arg-resolvers {'loopback-handler :injection}}
                                         ["fred" 'fred]
                                         ["barney" 'barney]
@@ -50,8 +32,8 @@
                                         ["creator-loopback" 'creator-loopback])
                 async/wrap-session
                 async/wrap-with-standard-middleware
-                (rook/wrap-with-injection :loopback-handler loopback-handler)
-                (jetty/run-async-jetty {:port 9988 :join? false :async-timeout 100}))))
+                (rook/wrap-with-injection :loopback-handler @handler)
+                (jetty/run-async-jetty {:port 9988 :join? false :async-timeout 100})))
 
   (it "did initialize the server successfully"
       (should-not-be-nil @server))